Move GetAppHash() to kube package
